The clang cells demand the cast gcc forgives (23-5 cure)
fabsf into a double is an implicit promotion; four clang cells
refused what every gcc cell waved through, and the local precheck
runs gcc. The cast is now explicit and the lesson is now a scar:
clang reviews every test before it ships, or CI does it the hard
way.

+++ test/test_drivetrain.c
@@ -555,7 +555,7 @@ static void TestTankSteer(void)
         }
         if (run == 0)
         {
-            yawSpin = fabsf(m3Body_GetAngularVelocity(hull).y);
+            yawSpin = (double)fabsf(m3Body_GetAngularVelocity(hull).y);
         }
         hashes[run] = m3World_Hash(world);
         m3DestroyWorld(world);
